[feature #28796915] Can we integrate OpenJPA L2 Cache with EHCache by first adding basic dependencies? Well, this is proving to be harder than we thought. Note the following,
We get a class cast exception  org.apache.openjpa.persistence.PersistenceProviderImpl cannot be cast to javax.persistence.spi.PersistenceProvider: org.apache.openjpa.persistence.PersistenceProviderImpl cannot be cast to javax.persistence.spi.PersistenceProvider if we enable the two commented lines in the current pom.xml. This is most probably due to openjpa dependency being added. However, removal of this was still leaving the exception.

Note that, on this working build, the ROOT/WEB-INF/lib does not have an openjpa jar in it.NO JAR! However, we did this before and it still gave the above exception, something must've changed. This shows how fragile the nature of these dependencies are.

Multiple classloaders loading openjpa seems to be the issue since openEJB also has a dependency in it. Right now, we rely on that and do not use our own jar.

Finally, CrudService.java was made non-final to prevent an NPE occuring due to the container not finding the attribute entityManager, which is PRESENT.