[feature #28796915] Can we integrate OpenJPA L2 Cache with EHCache by first adding basic dependencies? Well, this is proving to be harder than we thought. Note the following,
We get a class cast exception  org.apache.openjpa.persistence.PersistenceProviderImpl cannot be cast to javax.persistence.spi.PersistenceProvider: org.apache.openjpa.persistence.PersistenceProviderImpl cannot be cast to javax.persistence.spi.PersistenceProvider if we enable the two commented lines in the current pom.xml. This is most probably due to openjpa dependency being added. However, removal of this was still leaving the exception.

Note that, on this working build, the ROOT/WEB-INF/lib does not have an openjpa jar in it.NO JAR! However, we did this before and it still gave the above exception, something must've changed. This shows how fragile the nature of these dependencies are.

Multiple classloaders loading openjpa seems to be the issue since openEJB also has a dependency in it. Right now, we rely on that and do not use our own jar.

NEXT:

We include the ehcache depndency. Note that openjpa inclusion is specifically overridden by stating "provided". This too, works. Fragile!

Finally, CrudService.java was made non-final to prevent an NPE occuring due to the container not finding the attribute entityManager, which is PRESENT.

NEXT:

pom.xml and presistence.xml updated with ehcache details, the project works under very fragine conditions only. ehcache.xml should be placed int TOMCAT_HOME/lib and has to oveflow the stack to disk. If not, we get a  java.lang.StackOverflowError: null error upon consecutive refreshes on page. The project woked with tomcat under 400xm s and x, and the usual is 200.