304913-30602019 [feature] Converted Location Pages To Schema.Org Format into widget. We need to set properties on the widget now.
On a reporting basis, in case the following error occurs, make sure the xhtml content is validated. For example, missing closing quotes or unclosed tags. In this case it was a missing quote on an id tag which gave the following error. (There was nothing in logs).

Error 330 (net::ERR_CONTENT_DECODING_FAILED): Unknown error. (Chrome)

Firefox shows a similar but a bit confusing message complaining about compression.