[feature #37018575] I'm in the process of adding dynamic minification of JS files using YUI Compressor Maven Mojo. This means we need to move miniquarr out of  js in order to prevent the compressor from attempting to compress already compressed files (and create a new subdirectory under it called miniquarry). Compression is not tested yet. The home page also has some tweaks. Ok key down on the search box, the map opens up. However we're yet to be able to focus on the text input on the map, to give a continues sense of typing (as you see on Google these days). Calling #.focus() or #.click() isn't working. This also attempts to fix the focus body on hovering over it. This is ok, but the constant snapping is going to be removed (back to top snapping).
First focus on maps didn't work in Singapore. Street level searches led to zoom in to Satellite Mode, which was confusing. Now that happens at level 19 instead of 18.

I fixed home page images.
The support wording needs to be changed.
Also the home page SEARCH button is not apparent to be clickable and is easily confused with the text "I'm in".
Clicking on the text area (so that suggestion gets generated) is not very straight forward since it is already focused. Maybe the website should detect the location upon starting to type.

[Yes these are  the 3 previous commit messages. By committing the wrong message format Pivotal Tracker submissions didn't happen]
~How do I find exciting events in my area?