[feature #43107371] So we are now quite close to delivering location based events via email! Let's see what is there in this commit:
First, pom.xml has an upgraded version to cater for the new Geohash Subscribers
Helios is the God of Sun, or Sun himself. We now have a startup bean which will run timed events for us, and that is Helios
logback.xml runs in debug mode now until we get things more consistent.
We have a SubscriberNotifications.java which will timeout every 5 seconds to deliver the event emails. Mail delivery is not in place. Timeout needs to be fine tuned. Scanning happens partially on Stargate. This needs to be completed too. For now, we are going to go ahead with individual entry scanning. Next release we will focus on partial row scanning.
Oh, we also now have a small HBaseCrudService.java utility which will use the Stargate api to do CRUD operations. Maybe this utility should go in a new project on Github so that others can use. It will also simplify the main codebase.
That's about it for this commit
~Subscribe to events in this area in home page with HBase support